+++ src/messages/messageWebView.ts
@@ -9,7 +9,8 @@ export class MessageWebView {
     private panel: vscode.WebviewPanel | undefined;
 
     constructor(
-        private client: IServiceBusClient) {
+        private client: IServiceBusClient,
+        public readonly node: Subscription) {
     }
 
     async getMessages(topic: string, subscription: string, searchArguments: string | null): Promise<ReceivedMessageInfo[]> {
@@ -32,6 +33,12 @@ export class MessageWebView {
                         </td>
                         <td data-content-type="${x.contentType || ''}">
                             ${x.contentType || ''}
+                        </td>
+                        <td data-content-type="${x.label || ''}">
+                            ${x.label || ''}
+                        </td>
+                        <td data-content-type="${x.enqueuedSequenceNumber || ''}">
+                            ${x.enqueuedSequenceNumber || ''}
                         </td>
                         <td>
                             ${ x.enqueuedTimeUtc.toLocaleString() || ''}
@@ -61,7 +68,7 @@ export class MessageWebView {
             <head>
                 <meta charset="UTF-8">
                 <meta name="viewport" content="width=device-width, initial-scale=1.0">
-                <title>Cat Coding</title>
+                <title>Message List</title>
                 <style>
 
                     input {
@@ -142,6 +149,12 @@ export class MessageWebView {
                                     Content Type
                                 </th>
                                 <th style="text-align:left">
+                                    Label
+                                </th>
+                                <th style="text-align:left">
+                                    Enqueued Sequencenumber
+                                </th>
+                                <th style="text-align:left">
                                     TimeStamp
                                 </th>
                                 <th>
@@ -171,13 +184,13 @@ export class MessageWebView {
 
     }
 
-    async open(context: vscode.ExtensionContext, node: Subscription, searchArguments: string | null): Promise<void> {
+    async open(context: vscode.ExtensionContext, searchArguments: string | null): Promise<void> {
 
-        const messages = await this.getMessages(node.topicName, node.label, searchArguments);
+        const messages = await this.getMessages(this.node.topicName, this.node.label, searchArguments);
 
         this.panel = vscode.window.createWebviewPanel(
             'messagelist', // Identifies the type of the webview. Used internally
-            `${node.topicName} - (${node.label})`, // Title of the panel displayed to the user
+            `${this.node.topicName} - (${this.node.label})`, // Title of the panel displayed to the user
             vscode.ViewColumn.One, // Editor column to show the new webview panel in.
             {
                 enableScripts: true
@@ -197,7 +210,7 @@ export class MessageWebView {
             context.subscriptions
         );
 
-        await this.renderMessages(node.topicName, node.label, messages);
+        await this.renderMessages(this.node.topicName, this.node.label, messages);
 
         this.panel.onDidDispose(() => {